Adds config space read/write and improves type safety
Introduces typed access for reading and writing configuration space, enhancing mock transport fidelity and enabling safer config operations.
Updates type handling for DMA and BAR calculations to ensure correctness across address conversions.
Includes zerocopy dependency to support efficient, reliable byte conversions.

+++ drivers/kdriver/src/bus/pci.rs
@@ -76,7 +76,7 @@ fn config_pci_device<C: ConfigurationAccess>(
                         "  BAR {}: MEM [{:#x}, {:#x}){}{}",
                         bar,
                         address,
-                        address + size as u64,
+                        address + size,
                         if address_type == MemoryBarType::Width64 {
                             " 64bit"
                         } else {
